add vgg format
